modify the code structure to maven style
